fix(more-categories): keywords is not a list
Because keywords is not a list but a string with comma separated
values, every letter was separated. Instead, just take it as a
single item and lets append it as a single item, but if empty
append an empty list.

Also, for categories, when no category and no categories plugin
is found resulted in [None].